requirements_server.txt
# Para o servidor WebSocket
websockets

# Para a Detecção de Atividade de Voz (VAD)
webrtcvad-wheels

# Para a Transcrição (ElevenLabs)
elevenlabs

# Para a Análise (Grok-mini usa a biblioteca do OpenAI)
openai

# Para carregar as chaves de API do .env
python-dotenv